refactor : 컴포넌트 분리, 구조 리팩토링
